+++ docs/source/changelog.rst
@@ -35,6 +35,8 @@ Changelog
  - Support :class:`.Index` objects in :meth:`.DDLSession.create_table`.
 
  - Actually generate foreign keys upon table creation.
+ 
+ - Add Serial, BigSerial, and SmallSerial types to support automatic incrementation
 
 0.1.0 (released 2017-07-30)
 ---------------------------